Изменения для передачи параметра в детальный фрагмент погоды в городе

+++ app/src/main/java/online/tatarintsev/weather/model/entities/TownEntity.kt
@@ -6,6 +6,6 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class TownEntity (
     val name: String,
-    val lat: Double,
-    val lon: Double
+    val lat: Float,
+    val lon: Float
         ) : Parcelable
